// src/app/models/Permissions.ts
export enum Permissions {
    ASSIGN_BRANCH = 'Assign employees to OU',
    ASSIGN_PERMISSION = 'Permissions assigning',
    CREATE_CLIENT_ACCOUNT = 'Create clients account',
    CREATE_BRANCH = 'Organizationals unit create',
    CREATE_EMPLOYEE = 'Create employees',
    CREATE_MANAGER = 'Manager create',
    CREATE_REQUEST = 'Create request',
    CREATE_ROLE = 'Roles create',
    CREATE_USER = 'Users create',
    DELETE_BRANCH = 'Organizationals unit delete',
    DELETE_CLIENT_ACCOUNT = 'Delete clients account',
    DELETE_EMPLOYEE = 'Delete employees',
    DELETE_PERMISSION = 'Permissions delete from roles',
    DELETE_ROLE = 'Roles delete',
    DELETE_USER = 'Users delete',
    EDIT_REQUEST_STATUS = 'Change clients request status',
    EDIT_ROLE = 'Change users roles',
    REVIEW_BRANCH = 'Organizationals unit review',
    REVIEW_CLIENT_DATA = 'Clients data review',
    REVIEW_CLIENT_DOCUMENT = 'Clients documents review',
    REVIEW_CLIENT_REQUEST = 'Clients requests review',
    REVIEW_EMPLOYEE = 'Review employee data',
    REVIEW_MANAGER = 'Manager create', // TODO: add correct permission
    REVIEW_PERMISSIONS = 'Permissions review',
    REVIEW_ROLE = 'Roles review',
    REVIEW_USER = 'Users review',
    REVIEW_OWN_DOCUMENT ='Own documents review',
    REVIEW_OWN_REQUEST = 'Own requests review',
};